algo ok for very large stack (n > 500) and bench print in stderr

+++ benchmark.h
@@ -6,7 +6,7 @@
 /*   By: cvillene <cvillene@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/26 07:12:12 by cvillene          #+#    #+#             */
-/*   Updated: 2025/12/08 23:25:41 by cvillene         ###   ########.fr       */
+/*   Updated: 2025/12/15 23:39:53 by cvillene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,24 +16,24 @@
 
 typedef struct s_monitoring
 {
-	char	*strategy;
-	char	*time_order;
-	int		disorder;
-	int		n_sa;
-	int		n_sb;
-	int		n_ss;
-	int		n_pa;
-	int		n_pb;
-	int		n_ra;
-	int		n_rb;
-	int		n_rr;
-	int		n_rra;
-	int		n_rrb;
-	int		n_rrr;
+	char			*strategy;
+	char			*time_order;
+	unsigned long	disorder;
+	int				n_sa;
+	int				n_sb;
+	int				n_ss;
+	int				n_pa;
+	int				n_pb;
+	int				n_ra;
+	int				n_rb;
+	int				n_rr;
+	int				n_rra;
+	int				n_rrb;
+	int				n_rrr;
 }		t_monitoring;
 
 t_monitoring	add_monitoring(t_monitoring m1, t_monitoring m2);
-int				compute_disorder(t_stack *s);
+unsigned long	compute_disorder(t_stack *s);
 int				compute_total_ops(t_monitoring m);
 void			print_benchmark(t_monitoring m);
